introducing wayfinding.ipynb
